[refactor] 이슈 리스트 리팩토링
- 기존 이슈리스트 파일 분리
- 폴더 구조 정의에 맞추어 리팩토링
- 공용으로 사용하는 type 분리

- 이슈 리스트에 Issue 정보 표시
-
