+++ src/components/Issues/IssuesHeader.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Typography } from '@material-ui/core';
+
+function IssueHeader(): React.ReactElement {
+  return (
+    <Typography variant="h3" color="textPrimary">
+      Issues
+    </Typography>
+  );
+}
+
+export default IssueHeader;
